# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a ReID's server API.",
        "title": "ReID API",
        "termsOfService": "https://:).moe",
        "contact": {
            "name": "Kasper",
            "url": "https://kasper.moe",
            "email": "me@Kasper.moe"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/api",
    "paths": {
        "/query": {
            "post": {
                "description": "选取特定的行人",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "选取特定的行人",
                "operationId": "Query",
                "parameters": [
                    {
                        "type": "file",
                        "description": "行人图片",
                        "name": "files",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\": \"uploaded successfully\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"error\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "415": {
                        "description": "{\"error\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"error\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/search": {
            "post": {
                "description": "查找特定的行人",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "查找特定的行人",
                "operationId": "Search",
                "parameters": [
                    {
                        "type": "file",
                        "description": "待查找视频",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\": \"searched successfully\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"error\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "415": {
                        "description": "{\"error\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"error\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}